more robust and colorful git-cleanup function

+++ core.d/80_git.sh
@@ -6,23 +6,45 @@ if which -s git; then
   export SHELL_CONFIG_TOOLS="$SHELL_CONFIG_TOOLS git-cleanup"
 
   function git-cleanup {
-    git checkout develop
-    # verify that you want to !
+    YELLOW='\033[1;33m'
+    GREEN='\033[1;32m'
+    CLEAR='\033[00m'
 
-    echo "Fetching latest..."
+    echo -e "${YELLOW}Fetching latest${CLEAR}"
     git fetch
-    echo "Clearing stash..."
+
+    echo -e "${YELLOW}Resetting master${CLEAR}"
+    git checkout master > /dev/null
+    git reset --hard origin/master
+
+    if git branch | egrep '^. develop$' >/dev/null; then
+      echo -e "${YELLOW}Resetting develop${CLEAR}"
+      git checkout develop > /dev/null
+      git reset --hard origin/develop
+    fi
+
+    echo -e "${YELLOW}Clearing stash${CLEAR}"
     git stash clear
-    echo "Pruning removed branches..."
+
+    echo -e "${YELLOW}Pruning removed branches${CLEAR}"
     git remote prune origin
-    echo "Pruning merged branches..."
-    git branch --merged origin/develop | egrep -v '^. (develop|master)$' | xargs -n 1 git branch -v -d
-    echo "Pruning unused commits..."
+
+    echo -e "${YELLOW}Pruning merged branches${CLEAR}"
+    git branch --merged origin/master | egrep -v '^. (develop|master)$' | xargs -n 1 git branch -v -d
+
+    echo -e "${YELLOW}Pruning unused commits${CLEAR}"
     git prune
-    echo "GC..."
+
+    echo -e "${YELLOW}GC${CLEAR}"
     git gc
-    echo "Clearing logs..."
-    rm -fv log/*.log
-    echo "All clean :)"
+
+    if [ -d log ]; then
+      echo -e "${YELLOW}Clearing logs${CLEAR}"
+      rm -fv log/*.log
+    fi
+
+    echo ""
+    echo -e "${GREEN}All clean :)${CLEAR}"
+    echo ""
   }
 fi
